Modify/ PostingView, PostingDetailView with modified Posting model

+++ posts/views.py
@@ -11,8 +11,6 @@ from posts.serializers import (
     CommentCreateSerializer,
 )
 
-# Create your views here.
-
 
 # posting/
 class PostingView(APIView):
@@ -35,8 +33,7 @@ class PostingView(APIView):
     def post(self, request):
         serializer = PostingSerializer(data=request.data)
         if serializer.is_valid():
-            serializer.save()
-            # serializer.save(user=request.user)
+            serializer.save(user=request.user)
         else:
             return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
         return Response(serializer.data, status=status.HTTP_200_OK)
@@ -46,34 +43,32 @@ class PostingView(APIView):
 class PostingDetailView(APIView):
     """
     게시글 상세보기 페이지
-    id = posting_id인 게시글 1개 가져오기,
-    (댓글 추가시) 댓글 fk가 posting_id인 댓글들도 가져오기
+    id = posting_id인 게시글 1개 가져오기
     """
 
     def get(self, request, posting_id):
         posting = get_object_or_404(Posting, id=posting_id)
-        comment = Comment.objects.filter(id=posting_id)
+        # comment = Comment.objects.filter(id=posting_id)
         serializer = PostingDetailSerializer(posting)
         return Response(serializer.data, status=status.HTTP_200_OK)
 
     """
     게시글 수정하기
     posting.user == request.user인지 확인
-    get요청으로 받아온 값들 default로 채워주기
+    get요청으로 받아온 값들 default로 채워주기(프론트에서?)
     """
 
     def put(self, request, posting_id):
         posting = get_object_or_404(Posting, id=posting_id)
         serializer = PostingDetailSerializer(posting, data=request.data)
-        # if posting.user == request.user
-        if serializer.is_valid():
-            serializer.save()
-            # serializer.save(user=request.user)
-            return Response(serializer.data, status=status.HTTP_200_OK)
+        if posting.user == request.user:
+            if serializer.is_valid():
+                serializer.save(user=request.user)
+                return Response(serializer.data, status=status.HTTP_200_OK)
+            else:
+                return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
         else:
-            return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
-        # else:
-        #   return Response(status=status.HTTP_401_UNAUTHORIZED)
+            return Response(status=status.HTTP_401_UNAUTHORIZED)
 
     """
     게시글 삭제하기
@@ -82,11 +77,11 @@ class PostingDetailView(APIView):
 
     def delete(self, request, posting_id):
         posting = get_object_or_404(Posting, id=posting_id)
-        # if posting.user == request.user
-        posting.delete()
-        return Response(status=status.HTTP_204_NO_CONTENT)
-        # else:
-        #   return Response(status=status.HTTP_401_UNAUTHORIZED)
+        if posting.user == request.user:
+            posting.delete()
+            return Response(status=status.HTTP_204_NO_CONTENT)
+        else:
+            return Response(status=status.HTTP_401_UNAUTHORIZED)
 
 
 class CommentView(APIView):
